// src/uplink/interfaces/metadata/controller-metadata.interface.d.ts
import { BindingMetadata } from "./binding-metadata.interface";
import { EventMetadata } from "./event-metadata.interface";
import { MethodMetadata } from "./method-metadata.interface";
/**
 * Metadata for controller documentation and tooling
 */
export interface ControllerMetadata {
    name?: string;
    description?: string;
    bindings?: Record<string, BindingMetadata>;
    methods?: Record<string, MethodMetadata>;
    events?: Record<string, EventMetadata>;
}
//# sourceMappingURL=controller-metadata.interface.d.ts.map